feat(api): role on the wire, and contract examples taken from the server
The app track asked for shapes as examples rather than prose, having
already been burned by a documented auth body that did not match the
code. So the examples are now generated by driving the real app —
tools/contract_examples.py — and regenerating is the check that doc and
code still agree.

That immediately found a defect no test had: profiles.color was INTEGER,
and an Android colour is unsigned ARGB. 0xFF2A9D8F is 4283215696, past
the top of int32, stored happily on the device — so the first profile
carrying a colour would have failed its whole push. Synthetic zeros in
the tests hid it. Migration 0005 widens it.

Their ninth question was the important one, and it was right: a watcher
received a profile with owner_device_id projected away and could not
tell it from one of its own that nobody had claimed, so it would have
armed alarms for someone else's doses. The caller's role now travels on
the profile row — not profile_memberships, which would tell a watcher who
else can see it. Role answers the only question a device has to ask, and
the resulting rule has no ambiguous branch.

Also from their list: a resend is now a true no-op, since the upsert
requires a strictly newer updated_at — with >= it rewrote identical
values, took a new server_seq and pushed the row to every other device.
And a batch over 1000 records is refused whole rather than truncated,
because a truncated batch leaves the client believing it sent everything.

65 tests.

+++ api/app/db/models.py
@@ -141,7 +141,12 @@ class Profile(Base):
     )
 
     name: Mapped[str] = mapped_column(EncryptedString)
-    color: Mapped[int] = mapped_column(Integer, default=0)
+
+    # BIGINT, not INTEGER. An Android colour is unsigned ARGB — 0xFF2A9D8F is
+    # 4283215696, past the top of int32 — and SQLite stores it happily on the
+    # device. Synthetic zeros in the tests hid this; the first profile with a
+    # colour would have failed the whole push.
+    color: Mapped[int] = mapped_column(BigInteger, default=0)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
 
     # is_elder_mode is deliberately absent. It is a property of a device that
